repair UI screen controller in Administrator  & Remove addd new screen

+++ app/controllers/ScheduleData.php
@@ -57,15 +57,34 @@ class ScheduleData extends CI_Controller
         if(!isset($id) && !isset($site))
             redirect(base_url('site/list'));
 
-        $flight     = Schedule::find($id);        
-
-        if($flight->delete())
-            if($this->session->userdata('auth_role')== "Administrator")
-                redirect(base_url('site/show/').$site);
+        if($this->session->userdata('auth_role')!= "Administrator")
+        {
+            if($site != $this->session->userdata('auth_status'))
+            {
+                redirect(base_url('site/list'));  
+            }
             else
-                redirect(base_url());
-        else
-            redirect(base_url('site/show/').$site);
+            {
+                $flight     = Schedule::find($id);        
+
+                if($flight->delete())                    
+                    redirect(base_url());
+                else
+                    redirect(base_url());
+            }          
+        }
+        else 
+        {
+            $flight     = Schedule::find($id);        
+
+            if($flight->delete())
+                if($this->session->userdata('auth_role')== "Administrator")
+                    redirect(base_url('site/show/').$site);
+                else
+                    redirect(base_url());
+            else
+                redirect(base_url('site/show/').$site);
+        }        
     }
 }
